refactor test filter
