+++ test/integration/filters/header_to_filter_state.proto
@@ -1,0 +1,8 @@
+syntax = "proto3";
+
+package test.integration.filters;
+
+message HeaderToFilterStateFilterConfig {
+  string header_name = 1;
+  string state_name = 2;
+}
